Fix no button behavior on click text span in button
Addressed behavior by handling case in which button is the target
and case in which the span is the target, in which case the parent
(button) is referred to. Span reference identified using match function
and a selector for the button's class or id.

Poor commit, but also addressed issue of erroneous output not disabling
evaluation and operator buttons; set disabled for those buttons to true
instead of false.