Fix handling of numbers with more than 8 digits
Use regex grouping to identify digits before and after the decimal
point. If whole digits are greater than the max number of digits, error
out. Otherwise, round the digits after the decimal place to fit the
result on the screen while maintaining the whole number in the operand
variable.

Also apply number formatting to operand 1 if it is displayed by itself.